requirements.txt
fastapi>=0.110.0
uvicorn>=0.29.0
requests>=2.31.0
Pillow>=10.0.0
pydantic>=2.0.0